+++ Cargo.toml
@@ -42,6 +42,8 @@ serde_derive_state = { version = "0.4.0", optional = true }
 
 # Binding crates
 regex = { version = "1", optional = true }
+http = { version = "0.1", optional = true }
+hyper = { version = "0.12", optional = true }
 
 # Crates used in testing
 compiletest_rs = { version = "0.3", optional = true }
@@ -69,9 +71,6 @@ walkdir = "2"
 failure = "0.1"
 failure_derive = "0.1"
 
-hyper = "0.12.0"
-http = "0.1.0"
-
 serde = "1.0.0"
 serde_derive = "1.0.0"
 serde_derive_state = { version = "0.4.0" }
@@ -86,10 +85,11 @@ gluon_codegen = { path = "codegen", version = "0.9.1" } # GLUON
 [features]
 default = ["regex", "rand"]
 serialization = ["serde", "serde_state", "serde_derive_state", "gluon_vm/serialization"]
+web = ["hyper", "http"]
 
 docs_rs = ["serialization"]
 
-test = ["serialization", "little-skeptic", "gluon_vm/test", "gluon_check/test", "gluon_parser/test"]
+test = ["serialization", "little-skeptic", "http", "web", "gluon_vm/test", "gluon_check/test", "gluon_parser/test"]
 nightly = ["compiletest_rs"]
 test_nightly = ["test", "nightly"]
 
@@ -161,7 +161,7 @@ required-features = ["serialization"]
 [[example]]
 name = "http"
 path = "examples/http/main.rs"
-required-features = ["serialization"]
+required-features = ["serialization", "web"]
 [[example]]
 name = "lisp"
 path = "examples/lisp/main.rs"
