+++ examples/http/server.glu
@@ -29,7 +29,7 @@ let {
     method,
 
     ?
-} = import! examples.http.http
+} = import! std.http.http
 
 let hello_world : Handler Response =
     write_response (string.as_bytes "Hello World")
